refactor: moved DBEncryption to a separate class
