+++ packages/plugins/rudder_plugin_interface/lib/src/models/rudder_mobile_config.dart
@@ -1,3 +1,4 @@
+import 'db_encryption.dart';
 import '../constants.dart';
 
 ///Configurations to be used for mobile and mobile only. Any values set here
@@ -52,23 +53,4 @@ class MobileConfig {
   int get configRefreshInterval => _configRefreshInterval;
 
   DBEncryption? get dbEncryption => _dbEncryption;
-}
-
-class DBEncryption {
-  // @param whether the SDK should encrypt the DB
-  final bool _enabled;
-
-  // @param key to be used for encryption
-  final String _key;
-
-  DBEncryption(enabled, key)
-      : _enabled = enabled,
-        _key = key;
-
-  bool get enabled => _enabled;
-  String get key => _key;
-
-  Map getMap() {
-    return {"enabled": _enabled, "key": _key};
-  }
 }
